+++ tikv-client/src/main/java/com/pingcap/tikv/codec/CodecDataInput.java
@@ -1,107 +1,40 @@
-/*
- * Copyright 2017 PingCAP, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.pingcap.tikv.codec;
 
 import com.google.protobuf.ByteString;
+import gnu.trove.list.array.TByteArrayList;
+import io.netty.handler.codec.CodecException;
+import java.lang.reflect.Field;
+import java.nio.ByteOrder;
+import sun.misc.Unsafe;
 
-import java.io.*;
+public class CodecDataInput {
+  private static final Unsafe UNSAFE;
+  private static final int BYTE_ARRAY_BASE_OFFSET;
+  private static final boolean IS_LITTLE_ENDIAN;
 
-public class CodecDataInput implements DataInput {
-  /**
-   * An copy of ByteArrayInputStream without synchronization for faster decode.
-   *
-   * @see ByteArrayInputStream
-   */
-  private class UnSyncByteArrayInputStream extends InputStream {
-    protected byte buf[];
-    protected int pos;
-    protected int mark = 0;
-    protected int count;
-
-    UnSyncByteArrayInputStream(byte buf[]) {
-      this.buf = buf;
-      this.pos = 0;
-      this.count = buf.length;
+  static {
+    try {
+      Field field = Unsafe.class.getDeclaredField("theUnsafe");
+      field.setAccessible(true);
+      UNSAFE = (Unsafe)field.get(null);
+      BYTE_ARRAY_BASE_OFFSET = UNSAFE.arrayBaseOffset(byte[].class);
+      IS_LITTLE_ENDIAN = ByteOrder.nativeOrder().equals(ByteOrder.LITTLE_ENDIAN);
     }
-
-    public UnSyncByteArrayInputStream(byte buf[], int offset, int length) {
-      this.buf = buf;
-      this.pos = offset;
-      this.count = Math.min(offset + length, buf.length);
-      this.mark = offset;
-    }
-
-    public int read() {
-      return (pos < count) ? (buf[pos++] & 0xff) : -1;
-    }
-
-    public int read(byte b[], int off, int len) {
-      if (b == null) {
-        throw new NullPointerException();
-      } else if (off < 0 || len < 0 || len > b.length - off) {
-        throw new IndexOutOfBoundsException();
-      }
-
-      if (pos >= count) {
-        return -1;
-      }
-
-      int avail = count - pos;
-      if (len > avail) {
-        len = avail;
-      }
-      if (len <= 0) {
-        return 0;
-      }
-      System.arraycopy(buf, pos, b, off, len);
-      pos += len;
-      return len;
-    }
-
-    public long skip(long n) {
-      long k = count - pos;
-      if (n < k) {
-        k = n < 0 ? 0 : n;
-      }
-
-      pos += k;
-      return k;
-    }
-
-    public int available() {
-      return count - pos;
-    }
-    public boolean markSupported() {
-      return true;
-    }
-
-    public void mark(int readAheadLimit) {
-      mark = pos;
-    }
-
-    public void reset() {
-      pos = mark;
-    }
-
-    public void close() throws IOException {
+    catch (Exception e) {
+      throw new RuntimeException(e);
     }
   }
-  private final DataInputStream inputStream;
-  private final UnSyncByteArrayInputStream backingStream;
+
   private final byte[] backingBuffer;
+  private int pos;
+  private final int count;
+  private int mark;
+
+  private void checkLength(int required) {
+    if (backingBuffer.length - pos < required) {
+      throw new CodecException("EOF reached");
+    }
+  }
 
   public CodecDataInput(ByteString data) {
     this(data.toByteArray());
@@ -109,210 +42,138 @@ public class CodecDataInput implements DataInput {
 
   public CodecDataInput(byte[] buf) {
     backingBuffer = buf;
-    // MyDecimal usually will consume more bytes. If this happened,
-    // we need have a mechanism to reset backingStream.
-    // User mark first and then reset it later can do the trick.
-    backingStream =
-        new UnSyncByteArrayInputStream(buf) {
-          @Override
-          public void mark(int givenPos) {
-            mark = givenPos;
-          }
-        };
-    inputStream = new DataInputStream(backingStream);
+    count = backingBuffer.length;
+    mark = 0;
   }
 
-  @Override
-  public void readFully(byte[] b) {
-    try {
-      inputStream.readFully(b);
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
+  void readFully(byte[] b, int len) {
+    System.arraycopy(backingBuffer, pos, b, 0, len);
+    pos += len;
   }
 
-  @Override
-  public void readFully(byte[] b, int off, int len) {
-    try {
-      inputStream.readFully(b, off, len);
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
+  public void skipBytes(int n) {
+    pos += n;
   }
 
-  @Override
-  public int skipBytes(int n) {
-    try {
-      return inputStream.skipBytes(n);
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public boolean readBoolean() {
-    try {
-      return inputStream.readBoolean();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public byte readByte() {
-    try {
-      return inputStream.readByte();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public int readUnsignedByte() {
-    try {
-      return inputStream.readUnsignedByte();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public short readShort() {
-    try {
-      return inputStream.readShort();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public int readUnsignedShort() {
-    try {
-      return inputStream.readUnsignedShort();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public int readPartialUnsignedShort() {
-    try {
-      byte readBuffer[] = new byte[2];
-      inputStream.read(readBuffer, 0, 2);
-      return ((readBuffer[0] & 0xff) << 8) + ((readBuffer[1] & 0xff) << 0);
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public char readChar() {
-    try {
-      return inputStream.readChar();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public int readInt() {
-    try {
-      return inputStream.readInt();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public long readLong() {
-    try {
-      return inputStream.readLong();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public final long readPartialLong() {
-    try {
-      byte readBuffer[] = new byte[8];
-      inputStream.read(readBuffer, 0, 8);
-      return (((long) readBuffer[0] << 56) +
-          ((long) (readBuffer[1] & 255) << 48) +
-          ((long) (readBuffer[2] & 255) << 40) +
-          ((long) (readBuffer[3] & 255) << 32) +
-          ((long) (readBuffer[4] & 255) << 24) +
-          ((readBuffer[5] & 255) << 16) +
-          ((readBuffer[6] & 255) << 8) +
-          ((readBuffer[7] & 255) << 0));
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public float readFloat() {
-    try {
-      return inputStream.readFloat();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public double readDouble() {
-    try {
-      return inputStream.readDouble();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public String readLine() {
-    try {
-      return inputStream.readLine();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Override
-  public String readUTF() {
-    try {
-      return inputStream.readUTF();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public int peekByte() {
-    mark(currentPos());
-    int b = readByte() & 0xFF;
-    reset();
+  byte readByte() {
+    byte b = backingBuffer[pos];
+    pos ++;
     return b;
   }
 
-  public int currentPos() {
-    return size() - available();
+  public int readUnsignedByte() {
+    int r = backingBuffer[pos] & 0xFF;
+    pos ++;
+    return r;
+  }
+
+  public int readPartialUnsignedShort() {
+    if (count > pos) {
+      short r = UNSAFE.getShort(backingBuffer, (long) pos + BYTE_ARRAY_BASE_OFFSET);
+      if (IS_LITTLE_ENDIAN) {
+        r = Short.reverseBytes(r);
+      }
+      pos += 2;
+      return r & 0xFFFF;
+    }
+
+    if (count - pos == 1) {
+      int r = (backingBuffer[pos] & 0xFF) << 8;
+      pos ++;
+      return r;
+    }
+
+    return 0;
+  }
+
+  long readLong() {
+    checkLength(8);
+    long r = UNSAFE.getLong(backingBuffer, (long)pos + BYTE_ARRAY_BASE_OFFSET);
+    if (IS_LITTLE_ENDIAN) {
+      r = Long.reverseBytes(r);
+    }
+    pos += 8;
+    return r;
+  }
+
+  final long readPartialLong() {
+    if (available() >= 8) {
+      return readLong();
+    }
+
+    int shift = 56;
+    long r = 0;
+    if (available() > 0) {
+      r += (long) backingBuffer[pos] << shift;
+      pos++;
+      shift -= 8;
+    }
+
+    while (available() > 0 && shift >= 0) {
+      r += (long) (backingBuffer[pos] & 0xFF) << shift;
+      pos++;
+      shift -= 8;
+    }
+    return r;
+  }
+
+  public static final int GRP_SIZE = 8;
+  public static final int MARKER = 0xFF;
+  public static final byte[] PADS = new byte[GRP_SIZE];
+
+  // readBytes decodes bytes which is encoded by EncodeBytes before,
+  // returns the leftover bytes and decoded value if no error.
+  public byte[] readBytes() {
+    TByteArrayList cdo = new TByteArrayList(Math.min(available(), 256));
+    while (true) {
+      int padCount;
+      int marker = backingBuffer[pos + GRP_SIZE] & 0xff;
+      int curPos = pos;
+      pos += GRP_SIZE + 1;
+      padCount = MARKER - marker;
+
+      if (padCount > GRP_SIZE) {
+        throw new IllegalArgumentException("Wrong padding count");
+      }
+      int realGroupSize = GRP_SIZE - padCount;
+      cdo.add(backingBuffer, curPos, realGroupSize);
+
+      if (padCount != 0) {
+        // Check validity of padding bytes.
+        for (int i = realGroupSize; i < GRP_SIZE; i++) {
+          if (backingBuffer[i + curPos] != 0) {
+            throw new IllegalArgumentException();
+          }
+        }
+        break;
+      }
+    }
+
+    return cdo.toArray();
+  }
+
+  public int peekByte() {
+    return backingBuffer[pos] & 0xFF;
   }
 
   public void mark(int givenPos) {
-    this.backingStream.mark(givenPos);
+    mark = givenPos;
   }
 
   public void reset() {
-    this.backingStream.reset();
+    pos = mark;
   }
 
   public boolean eof() {
-    return backingStream.available() == 0;
+    return available() == 0;
   }
 
   public int size() {
-    return backingBuffer.length;
+    return count;
   }
 
   public int available() {
-    return backingStream.available();
+    return count - pos;
   }
 
   public byte[] toByteArray() {
